+++ assets/version_check.py
@@ -16,7 +16,7 @@ CHANGELOG_ITEMS = [
 ISSUE_TEMPLATE_ITEMS = ["- SpinHook {0}"]
 SECURITY_ITEMS = ["| {0}           | :white_check_mark: |", "| < {0}         | :x:                |"]
 SERVICE_WORKER_ITEMS = ["spinhook-v{0}"]
-SCRIPT_ITEMS = ['GAME_VERSION = "v{0}"']
+SCRIPT_ITEMS = ['SPINHOOK_VERSION = "v{0}"']
 
 FILES = {
     #"index.html": INDEX_ITEMS,